api/requirements.txt
build==0.10.0
packaging==23.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
tomli==2.0.1

pip~=21.3.1
attrs~=22.2.0
wheel~=0.37.1
setuptools~=60.2.0
openai~=0.27.2